clean up the python API, separate the main from the API functions.
