+++ setup.py
@@ -16,7 +16,7 @@ setup(
         'docopt',
     ],
     entry_points={'console_scripts': [
-        'corsika = corsika_wrapper.__init__:main',
+        'corsika = corsika_wrapper.main:main',
     ]},
     zip_safe=False,
 )
